pyproject.toml
[project]
name = "login-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.4.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
